Move expanding cycles logic to a separate file
Otherwise, it looks like CheckedCSubst.cpp grows too large because
of all the visitors in it, leading to an error while generating the obj
file.
